refactor(spec): externalize plugin preflight test surface
